+++ client/app/login/page.tsx
@@ -1,0 +1,5 @@
+import Login from "../../src/views/Login";
+
+export default function LoginPage() {
+  return <Login />;
+}