+++ client/app/register/page.tsx
@@ -1,0 +1,5 @@
+import Register from "../../src/views/Register";
+
+export default function RegisterPage() {
+  return <Register />;
+}